requirements.txt
requests[socks]>=2.31.0
PySocks>=1.7.1
colorama>=0.4.6
stem>=1.8.2
tqdm>=4.66.1